XWIKI-14218: Add the ability to disable the notifications * Fix file formatting * Separate role from implementation
