+++ xwiki-platform-core/xwiki-platform-notifications/xwiki-platform-notifications-api/src/main/resources/META-INF/components.txt
@@ -1,5 +1,5 @@
-org.xwiki.notifications.NotificationConfiguration
 org.xwiki.notifications.internal.DefaultCompositeEventStatusManager
+org.xwiki.notifications.internal.DefaultNotificationConfiguration
 org.xwiki.notifications.internal.DefaultNotificationDisplayer
 org.xwiki.notifications.internal.DefaultNotificationManager
 org.xwiki.notifications.internal.DefaultNotificationRenderer
